customer_portal_mcp/server/requirements.txt
fastmcp>=1.0.0
python-dotenv>=1.0.1
pyyaml>=6.0.1
starlette>=0.27.0
uvicorn>=0.23.0
asyncpg>=0.29.0
